Move isLiteral and isEvaluated into ExpressionCodegen.
